+++ .env
@@ -1,0 +1,1 @@
+NEXT_PUBLIC_URL_API=https://ameco-development.up.railway.app/api
