+++ app/dashboard/affiliates/page.js
@@ -1,7 +1,8 @@
 'use client'
 
 import AffiliatesTable from '@/components/dashboard/affiliates-table/affiliates-table.js'
-import { useSelector } from 'react-redux/es/hooks/useSelector.js'
+import { formatDate,dateToNumber } from '@/app/functions'
+import { useSelector } from 'react-redux'
 import { useState,useEffect } from 'react'
 import { useRouter } from 'next/navigation'
 import axios from 'axios'
@@ -9,25 +10,7 @@ import axios from 'axios'
 export default function Affiliates(){
     const router=useRouter()
     const [data,setData]=useState('');
-    const user=useSelector(state=>state.user)
-
-    function formatDate(date){
-        const fechaStr = date;
-        const fechaObj = new Date(fechaStr);
-
-        const dia = fechaObj.getDate();
-        const mes = fechaObj.getMonth() + 1;
-        const año = fechaObj.getFullYear();
-
-        const fechaFormateada = `${dia}/${mes < 10 ? '0' + mes : mes}/${año}`;
-        return fechaFormateada;
-    }
-    function dateToNumber(fecha) {
-        var partes = fecha.split('/');
-        var fechaObjeto = new Date(partes[2], partes[1] - 1, partes[0]);
-        var numero = fechaObjeto.getTime();
-        return numero;
-    }
+    let user=useSelector(state=>state.user)
     
     useEffect(()=>{
         if (!user.user) {
@@ -54,7 +37,7 @@ export default function Affiliates(){
         }).catch(err=>{
             console.log(err)
         })
-    },[user])
+    },[])
 
     return (
         <div style={{"margin":"40px"}}>
