+++ app/dashboard/employers/page.js
@@ -1,0 +1,8 @@
+import React from 'react'
+
+export default function Employers() {
+  return (
+    <div>Employers</div>
+  )
+}
+
