+++ components/dashboard/affiliates-table/actions.js
@@ -5,13 +5,11 @@ import { useSelector } from 'react-redux'
 
 export default function Actions() {
   const user=useSelector(state=>state.user)
-  console.log('user',user.user.roles)
-  // const userRoles=user.user.roles
-  const userRoles=['recaudador']
+  const userRoles=user.user.roles
   return (
     <div>
     {
-      userRoles.includes('comision')||userRoles.includes('recaudador')?
+      userRoles.includes('comision') || userRoles.includes('recaudador')?
       <Dropdown options={['Ver ficha']}/>:
       <Dropdown options={['Otorgar beneficio', 'Ver ficha', 'Dar de baja']}/>
     }
